+++ ikii-back-commons/src/main/java/mx/ikii/commons/feignclient/service/impl/BusinessFeignServiceImpl.java
@@ -5,6 +5,8 @@ import org.springframework.stereotype.Service;
 
 import mx.ikii.commons.feignclient.repository.IBusinessFeignRepository;
 import mx.ikii.commons.mapper.business.IBusinessCategoryMapper;
+import mx.ikii.commons.mapper.business.IBusinessMapper;
+import mx.ikii.commons.persistence.collection.Business;
 import mx.ikii.commons.persistence.collection.BusinessCategory;
 import mx.ikii.commons.utils.ResponseEntityHelper;
 
@@ -16,11 +18,21 @@ public class BusinessFeignServiceImpl implements IBusinessFeignService {
 
 	@Autowired
 	private IBusinessCategoryMapper businessCategortyMapper;
+	
+	@Autowired
+	private IBusinessMapper businessMapper;
+	
 
 	@Override
 	public BusinessCategory getByCategoryId(String categoryId) {
 		return businessCategortyMapper.responseToEntity(
 				ResponseEntityHelper.processingHttpStatus(businessFeignRespository.getByCategoryId(categoryId)));
+	}
+
+	@Override
+	public Business getById(String id) {
+		return businessMapper.responseToEntity(
+				ResponseEntityHelper.processingHttpStatus(businessFeignRespository.getById(id)));
 	}
 
 }