+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/response/business/BusinessResponse.java
@@ -17,6 +17,7 @@ public class BusinessResponse {
 	private String closeTime;
 	private Boolean isOpen;
 	private String status;
+	private boolean isFavorite;
 	private String postalCode;
 
 }