+++ ikii-back-commons/src/main/java/mx/ikii/commons/persistence/collection/CustomerDetails.java
@@ -1,21 +1,29 @@
 package mx.ikii.commons.persistence.collection;
 
+import java.util.List;
+
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 @Document(collection = "CustomerDetails")
 @JsonInclude(Include.NON_NULL)
 public class CustomerDetails {
-
 	private ObjectId customerId;
 	private Boolean isValidAccount;
 	private String image;
 	private Integer type;
-	
+	private List<ObjectId> businessFavorites;
+
 }