+++ ikii-back-microservices/ikii-back-customers/src/main/java/mx/ikii/customers/repository/ICustomerDetailsRepository.java
@@ -2,6 +2,7 @@ package mx.ikii.customers.repository;
 
 import java.util.Optional;
 
+import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -16,5 +17,5 @@ import mx.ikii.commons.persistence.collection.CustomerDetails;
  */
 @Repository
 public interface ICustomerDetailsRepository extends MongoRepository<CustomerDetails, String> {
-	public Optional<CustomerDetails> findByCustomerId(String customerId);
+	public Optional<CustomerDetails> findByCustomerId(ObjectId customerId);
 }